requirements.txt
fastapi==0.129.2
pydantic==1.10.12
PyYAML==6.0.1
Requests==2.32.5
